+++ Sources/Utilities/ProfanityFilter.swift
@@ -13,14 +13,22 @@ struct ProfanityFilter {
         case fileNotFoundError
         case formattingError
     }
-    let fileName: String
+    let fileName: String?
+    let restrictedMessageRegex: String?
     var restrictedWords = Set<String>()
 
     let bundle: Bundle
 
-    init(fileName: String, bundle: Bundle = .main) throws {
+    private init(
+        fileName: String?,
+        messageRegex: String?,
+        bundle: Bundle) throws {
+
         self.fileName = fileName
+        self.restrictedMessageRegex = messageRegex
         self.bundle = bundle
+
+        guard let fileName = fileName else { return }
         do {
             restrictedWords = try restrictedWords(fileName: fileName)
         } catch {
@@ -50,8 +58,36 @@ struct ProfanityFilter {
         var isPresent = false
         for word in wordsInText {
             isPresent = restrictedWords.contains(word)
-            break
+            if isPresent { return true }
+        }
+
+        // If not present in the restricted words then match the text
+        // against the pattern.
+        if let restrictedMessagePattern = restrictedMessageRegex {
+            let range = NSRange(text.startIndex..<text.endIndex, in: text)
+            do {
+                let regex = try NSRegularExpression(pattern: restrictedMessagePattern, options: [])
+                let matches = regex.numberOfMatches(in: text, options: [], range: range)
+                print("Restricted text matches: \(matches)")
+                if matches > 0 { isPresent = true }
+            } catch {
+                print("Error while matching restricted text: \(error.localizedDescription)")
+            }
         }
         return isPresent
+    }
+}
+
+extension ProfanityFilter {
+    init(fileName: String, bundle: Bundle = .main) throws {
+        try self.init(fileName: fileName, messageRegex: nil, bundle: bundle)
+    }
+
+    init(restrictedMessageRegex: String, bundle: Bundle = .main) throws {
+        try self.init(fileName: nil, messageRegex: restrictedMessageRegex, bundle: bundle)
+    }
+
+    init(fileName: String, restrictedMessageRegex: String, bundle: Bundle = .main) throws {
+        try self.init(fileName: fileName, messageRegex: restrictedMessageRegex, bundle: bundle)
     }
 }